project/requirements.txt
Flask
line-bot-sdk==3.11.0
openai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
bleak
psutil
opencv-python
ultralytics==8.2.22
supervision==0.20.0
numpy
python-dotenv
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
ngrok
